# quality/openapi/media-files.json
{
  "components" : {
    "schemas" : {
...
      "FileBusinessError" : {
        "properties" : {
          "code" : {
            "description" : "业务错误码",
            "type" : "string"
          },
          "details" : {
            "additionalProperties" : { },
            "description" : "可选错误上下文；内容由错误码决定",
            "type" : "object"
          },
          "message" : {
            "description" : "错误说明",
            "type" : "string"
          },
          "path" : {
            "description" : "请求路径",
            "type" : "string"
          },
          "requestId" : {
            "description" : "请求追踪 ID",
            "type" : "string"
          },
          "retryable" : {
            "description" : "是否允许重试",
            "type" : "boolean"
          },
          "status" : {
            "description" : "HTTP 状态码",
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "description" : "上海时区错误时间",
            "type" : "string"
          }
        },
        "required" : [ "code", "message", "path", "requestId", "retryable", "status", "timestamp" ],
        "type" : "object"
      },
...
      "FileUploadProgressItem" : {
        "properties" : {
          "errorCode" : {
            "description" : "错误码；无错误时为空",
            "type" : [ "string", "null" ]
          },
          "errorMessage" : {
            "description" : "失败原因；无错误时为空",
            "type" : [ "string", "null" ]
          },
          "fileId" : {
            "description" : "文件 ID",
            "type" : "string"
          },
          "fileName" : {
            "description" : "原始文件名",
            "type" : "string"
          },
          "fileType" : {
            "description" : "文件业务类型",
            "type" : "string"
          },
          "partCount" : {
            "description" : "分片总数；没有上传会话时为空",
            "format" : "int32",
            "type" : [ "integer", "null" ]
          },
          "phase" : {
            "description" : "上传及后处理的统一展示阶段",
            "enum" : [ "UPLOADING", "FINALIZING", "PROCESSING", "READY", "FAILED", "EXPIRED", "ABORTED", "DELETED", "UNKNOWN" ],
            "type" : "string"
          },
          "ready" : {
            "description" : "文件及必要后处理是否已就绪",
            "type" : "boolean"
          },
          "totalBytes" : {
            "description" : "预期文件总字节数",
            "format" : "int64",
            "type" : "integer"
          },
          "updatedAt" : {
            "description" : "进度更新时间",
            "example" : "2026-09-29 12:00:00",
            "pattern" : "^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}:\\d{2}$",
            "type" : [ "string", "null" ]
          },
          "uploadId" : {
            "description" : "平台上传会话 ID；来源文件已完成时为空",
            "type" : [ "string", "null" ]
          },
          "uploadPercent" : {
            "description" : "上传百分比，100 不等于后处理已完成",
            "format" : "double",
            "type" : "number"
          },
          "uploadedBytes" : {
            "description" : "已确认上传的字节数",
            "format" : "int64",
            "type" : "integer"
          },
          "uploadedPartCount" : {
            "description" : "已确认上传分片数",
            "format" : "int32",
            "type" : "integer"
          },
          "version" : {
            "description" : "进度版本；用于识别进度更新，终态快照可为 0",
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "errorCode", "errorMessage", "fileId", "fileName", "fileType", "partCount", "phase", "ready", "totalBytes", "updatedAt", "uploadId", "uploadPercent", "uploadedBytes", "uploadedPartCount", "version" ],
        "type" : "object"
      },
      "FileUploadProgressQueryRequest" : {
        "properties" : {
          "fileIds" : {
            "description" : "查询文件 ID，最多 500 个；未传、null 或空数组返回空结果",
            "items" : {
              "type" : "string"
            },
            "maxItems" : 500,
            "minItems" : 0,
            "type" : [ "array", "null" ]
          }
        },
        "type" : "object"
      },
      "FileUploadProgressQueryResponse" : {
        "properties" : {
          "generatedAt" : {
            "description" : "本次进度查询生成时间",
            "example" : "2026-09-29 12:00:00",
            "pattern" : "^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}:\\d{2}$",
            "type" : "string"
          },
          "items" : {
            "description" : "本次返回的文件记录",
            "items" : {
              "$ref" : "#/components/schemas/FileUploadProgressItem"
            },
            "type" : "array"
          },
          "missingFileIds" : {
            "description" : "无法识别的文件 ID；与 items 分别返回",
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "generatedAt", "items", "missingFileIds" ],
        "type" : "object"
      },
...
      "FrameworkError" : {
        "properties" : {
          "error" : {
            "description" : "HTTP 错误名称",
            "type" : "string"
          },
          "path" : {
            "description" : "请求路径",
            "type" : "string"
          },
          "status" : {
            "description" : "HTTP 状态码",
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "description" : "错误时间",
            "type" : "string"
          }
        },
        "required" : [ "error", "path", "status", "timestamp" ],
        "type" : "object"
      },
      "MinioPartEvent" : {
        "properties" : {
          "Records" : {
            "description" : "对象事件数组；启用进度处理时必须提供",
            "items" : {
              "$ref" : "#/components/schemas/PartRecord"
            },
            "type" : "array"
          }
        },
        "required" : [ "Records" ],
        "type" : "object"
      },
      "PartObject" : {
        "description" : "本次事件涉及的对象键、字节数与 ETag",
        "properties" : {
          "eTag" : {
            "description" : "对象 ETag；缺失时按空字符串处理",
            "type" : "string"
          },
          "key" : {
            "description" : "URL 编码对象键；分片键包含 .upload-parts/{storageUploadId}/part-{number}",
            "type" : "string"
          },
          "size" : {
            "description" : "分片字节数；匹配的分片事件必须为非负值",
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "PartRecord" : {
        "properties" : {
          "eventName" : {
            "description" : "仅处理 s3:ObjectCreated: 前缀的事件",
            "type" : "string"
          },
          "s3" : {
            "$ref" : "#/components/schemas/StorageEvent",
            "description" : "对象信息"
          }
        },
        "type" : "object"
      },
...
      "StorageEvent" : {
        "description" : "对象存储通知中的对象信息",
        "properties" : {
          "object" : {
            "$ref" : "#/components/schemas/PartObject",
            "description" : "本次事件涉及的对象"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "MinioWebhookToken" : {
        "description" : "配置的内部回调凭证，使用常量时间比较；不是用户 JWT",
        "scheme" : "bearer",
        "type" : "http"
      },
      "PlaybackToken" : {
        "description" : "文件播放签名，绑定文件及有效期；不是用户 JWT",
        "in" : "query",
        "name" : "token",
        "type" : "apiKey"
      },
      "RobotId" : {
        "description" : "受控机器人标识，非独立签名凭据；公开路径还受来源网络配置限制",
        "in" : "header",
        "name" : "X-Robot-Id",
        "type" : "apiKey"
      }
    }
  },
  "info" : {
    "description" : "Media 仅供受控网络调用。用户接口信任上游身份 Header，解析器仍保留开发默认值；机器人分片接口依赖 X-Robot-Id 及公开路径的来源网络过滤；HLS 使用播放 token；MinIO 回调使用内部 Bearer 凭证。不同入口不能套用统一用户 JWT 鉴权。",
    "title" : "Media 文件接口",
    "version" : "1.0.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/internal/media/file-upload-events/minio" : {
      "post" : {
        "description" : "校验配置的内部 Bearer 凭证；识别 s3:ObjectCreated: 事件与分片对象键，忽略无关对象和未知会话；重复分片由进度服务去重。",
        "operationId" : "mediaReceiveFilePartEventInternal",
        "parameters" : [ {
          "description" : "MinIO 回调内部 Bearer 凭证；不是用户 JWT",
          "in" : "header",
          "name" : "Authorization",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MinioPartEvent"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "处理完成，无正文"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "oneOf" : [ {
                    "$ref" : "#/components/schemas/FileBusinessError"
                  }, {
                    "$ref" : "#/components/schemas/FrameworkError"
                  } ]
                }
              }
            },
            "description" : "请求校验或绑定失败"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "oneOf" : [ {
                    "$ref" : "#/components/schemas/FileBusinessError"
                  }, {
                    "$ref" : "#/components/schemas/FrameworkError"
                  } ]
                }
              }
            },
            "description" : "机器人标识或播放/内部凭证无效"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FrameworkError"
                }
              }
            },
            "description" : "未处理的服务错误"
          },
          "503" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "oneOf" : [ {
                    "$ref" : "#/components/schemas/FileBusinessError"
                  }, {
                    "$ref" : "#/components/schemas/FrameworkError"
                  } ]
                }
              }
            },
            "description" : "存储或内部凭证配置不可用"
          }
        },
        "security" : [ {
          "MinioWebhookToken" : [ ]
        } ],
        "summary" : "接收 MinIO 分片事件",
        "tags" : [ "files" ]
      }
    },
...
    "/internal/media/files/upload-progress-queries" : {
      "post" : {
        "description" : "仅受控内网调用，本入口没有 Bearer 凭证校验。最多 500 个 ID；去空白和重复后查询，缺失 ID 单独返回。百分比为 100 不代表后处理就绪。",
        "operationId" : "mediaQueryFileUploadProgressInternal",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FileUploadProgressQueryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FileUploadProgressQueryResponse"
                }
              }
            },
            "description" : "处理成功；状态、空值和部分失败以响应字段为准"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "oneOf" : [ {
                    "$ref" : "#/components/schemas/FileBusinessError"
                  }, {
                    "$ref" : "#/components/schemas/FrameworkError"
                  } ]
                }
              }
            },
            "description" : "请求校验或绑定失败"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FrameworkError"
                }
              }
            },
            "description" : "未处理的服务错误"
          }
        },
        "summary" : "批量查询上传进度",
        "tags" : [ "files" ]
      }
    },
...
  },
  "security" : [ ],
  "servers" : [ {
    "url" : "/"
  } ],
...
}